+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: yooshima <yooshima@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/06 16:45:28 by yooshima          #+#    #+#             */
-/*   Updated: 2024/05/08 16:19:56 by yooshima         ###   ########.fr       */
+/*   Updated: 2024/05/09 12:22:02 by yooshima         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int		put_hex(unsigned long long nb, int upper);
 int		put_pointer(unsigned long long p);
 int		put_str(char *s);
 int		put_c(int c);
-int		put_unsign_int(long nb);
+int		put_unsign_int(unsigned int nb);
 int		put_nbr(int n);
 
 #endif
